Refactor to run API-calling as server actions.
This enables API-calling code to have access to internal network
and inherit AWS permissions from server profile, without needing to
make those resources publicly accessible or store credentials on clients.
