+++ webui/src/app/types.ts
@@ -1,0 +1,5 @@
+export interface jobType {
+    uuid?: string,
+    status: string,
+    inputValidationPassed?: boolean
+}
